% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_summary.R
\name{break_func}
\alias{break_func}
\title{Take a statcast data set and visualize horizontal and vertical break by pitch type}
\usage{
break_func(file = NULL)
}
\arguments{
\item{file}{statcast file for specified player}
}
\value{
ggplot graphic
}
\description{
Take a statcast data set and visualize horizontal and vertical break by pitch type
}
\examples{
break_func(file = deGrom)
}
