% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_location_heatmap.R
\name{contact_location_heatmap}
\alias{contact_location_heatmap}
\title{Take a statcast data set and created a heatmap of hit pitch locations}
\usage{
contact_location_heatmap(
  file = NULL,
  by_pitch_type = TRUE,
  hit.status = FALSE,
  min.exit.velocity = 0
)
}
\arguments{
\item{file}{statcast file}

\item{by_pitch_type}{boolean to facet by pitch type, default is TRUE}

\item{hit.status}{boolean to filter contacts into only events resulting in hits}

\item{min.exit.velocity}{double to define the minimum exit velocity on a contact that you want to consider in the graphic}
}
\value{
ggplot graphic
}
\description{
Take a statcast data set and created a heatmap of hit pitch locations
}
\examples{
example1 <- contact_location_heatmap(file = deGrom, by_pitch_type = TRUE, hit.status = FALSE, min.exit.velocity =0.0)
example1
example2 <- contact_location_heatmap(file = deGrom, by_pitch_type = TRUE, hit.status = TRUE, min.exit.velocity =90.0)
example2

}
