% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_heatmap.R
\name{custom_heatmap}
\alias{custom_heatmap}
\title{Take a statcast data set and created a heatmap of pitch locations that are specific to the strikezone (always 3x3 in the zone)}
\usage{
custom_heatmap(file = NULL)
}
\arguments{
\item{file}{statcast file}
}
\value{
ggplot graphic
}
\description{
Take a statcast data set and created a heatmap of pitch locations that are specific to the strikezone (always 3x3 in the zone)
}
