% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_heatmap.R
\name{location_heatmap}
\alias{location_heatmap}
\title{Take a statcast data set and created a heatmap of pitch locations}
\usage{
location_heatmap(file = NULL, by_pitch_type = TRUE)
}
\arguments{
\item{file}{statcast data file}

\item{by_pitch_type}{boolean to facet by pitch type, default is TRUE}
}
\value{
ggplot graphic
}
\description{
Take a statcast data set and created a heatmap of pitch locations
}
