% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swg_strike.R
\name{swg_strike}
\alias{swg_strike}
\title{Take a statcast data set and acquire the swinging strike percentage}
\usage{
swg_strike(file)
}
\arguments{
\item{file}{statcast file}
}
\value{
summary table
}
\description{
Take a statcast data set and acquire the swinging strike percentage
}
\examples{
swg_strike(deGrom)
}
