% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo_time.R
\name{velo_time}
\alias{velo_time}
\title{Take a statcast data set and plot aver velocity per game of a certai pitch type}
\usage{
velo_time(file, pitch = "FF")
}
\arguments{
\item{file}{statcast file}

\item{pitch}{pitch type, default is FF}
}
\value{
summary table
}
\description{
Take a statcast data set and plot aver velocity per game of a certai pitch type
}
\examples{
velo_time(deGrom)
}
