% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_heatmap.R
\name{custom_heatmap}
\alias{custom_heatmap}
\title{Take a statcast data set and created a heatmap of pitch locations}
\usage{
custom_heatmap(file = NULL, by_pitch_type = TRUE, sq = 3)
}
\arguments{
\item{file}{statcast file}

\item{by_pitch_type}{boolean to facet by pitch type, default is TRUE}
}
\value{
ggplot graphic
}
\description{
Take a statcast data set and created a heatmap of pitch locations
}
